seperate member functionality
