+++ templates/base.html
@@ -31,8 +31,8 @@
                 <ul class="navbar-nav me-auto">
                     <li class="nav-item"><a class="nav-link" href="/books">Books</a></li>
                     <li class="nav-item"><a class="nav-link" href="/new-book">Add a New Book</a></li>
-                    <li class="nav-item"><a class="nav-link" href="{{ url_for('new_member') }}">Add Member</a></li>
-                    <li class="nav-item"><a class="nav-link" href="/list-member">View Existing Member</a></li>
+                    <li class="nav-item"><a class="nav-link" href="{{ url_for('members.new_member') }}">Add Member</a></li>
+                    <li class="nav-item"><a class="nav-link" href="{{ url_for('members.list_member') }}">View Existing Member</a></li>
                     <li class="nav-item"><a class="nav-link" href="/issue-book">Issue Book</a></li>
                     <li class="nav-item"><a class="nav-link" href="/list-books">List of Books</a></li>
                     <li class="nav-item"><a class="nav-link" href="/issued-books">Issued Books</a></li>
