+++ templates/memberTemplates/edit-member.html
@@ -6,7 +6,7 @@
         Edit Member
     </h2>
     <div class="card p-4 shadow-lg" style="background: #f5e3c3; border-radius: 15px; max-height: 80vh; overflow-y: auto;">
-        <form method="POST" action="{{ url_for('edit_member', member_id=member.member_id) }}">
+        <form method="POST" action="{{ url_for('members.edit_member', member_id=member.member_id) }}">
             <div class="mb-3">
                 <label for="first_name" class="form-label">First Name</label>
                 <input type="text" class="form-control" id="first_name" name="first_name" value="{{ member.first_name }}" required>
@@ -29,7 +29,7 @@
             </div>
             <div class="d-flex justify-content-between">
                 <button type="submit" class="btn btn-dark" style="background: #8b5e3c; border: none;">Save Changes</button>
-                <a href="{{ url_for('list_member') }}" class="btn btn-secondary">Cancel</a>
+                <a href="{{ url_for('members.list_member') }}" class="btn btn-secondary">Cancel</a>
             </div>
         </form>
     </div>
