+++ templates/memberTemplates/list-member.html
@@ -32,8 +32,8 @@
                     <td>{{ member.email }}</td>
                     <td>{{ member.phone }}</td>
                     <td>
-                        <a href="{{ url_for('edit_member', member_id=member.member_id) }}" class="btn btn-sm btn-primary">Edit</a>
-                        <form action="{{ url_for('delete_member', member_id=member.member_id) }}" method="POST" style="display:inline;">
+                        <a href="{{ url_for('members.edit_member', member_id=member.member_id) }}" class="btn btn-sm btn-primary">Edit</a>
+                        <form action="{{ url_for('members.delete_member', member_id=member.member_id) }}" method="POST" style="display:inline;">
                             <button type="submit" class="btn btn-sm btn-danger" onclick="return confirm('Are you sure you want to delete this member?')">
                                 Delete
                             </button>
@@ -50,15 +50,15 @@
     <nav aria-label="Page navigation" class="mt-3">
         <ul class="pagination justify-content-center">
             {% if pagination.prev_page %}
-                <li class="page-item"><a class="page-link" href="{{ url_for('list_member', page=pagination.prev_page) }}">Previous</a></li>
+                <li class="page-item"><a class="page-link" href="{{ url_for('members.list_member', page=pagination.prev_page) }}">Previous</a></li>
             {% endif %}
             {% for page_num in pagination.page_range %}
                 <li class="page-item {% if page_num == pagination.current_page %}active{% endif %}">
-                    <a class="page-link" href="{{ url_for('list_member', page=page_num) }}">{{ page_num }}</a>
+                    <a class="page-link" href="{{ url_for('members.list_member', page=page_num) }}">{{ page_num }}</a>
                 </li>
             {% endfor %}
             {% if pagination.next_page %}
-                <li class="page-item"><a class="page-link" href="{{ url_for('list_member', page=pagination.next_page) }}">Next</a></li>
+                <li class="page-item"><a class="page-link" href="{{ url_for('members.list_member', page=pagination.next_page) }}">Next</a></li>
             {% endif %}
         </ul>
     </nav>
