+++ templates/memberTemplates/new-member.html
@@ -5,7 +5,7 @@
 <div class="container mt-5 mb-5">
     <h2 class="text-center mb-4" style="color: #5a3e2b; font-family: 'Playfair Display', serif;">Add New Member</h2>
     <div class="card p-4 shadow-lg" style="background: #f5e3c3; border-radius: 15px; max-height: 80vh; overflow-y: auto;">
-        <form method="POST" action="/new-member">
+        <form method="POST" action="/members/new-member">
             <div class="mb-3">
                 <label for="first_name" class="form-label">First Name</label>
                 <input type="text" class="form-control" id="first_name" name="first_name" required>
